type: feat(package): add `embla-carousel-react` and update `@types/node`
This commit introduces a new dependency, `embla-carousel-react` at version `8.0.0-rc15`, for carousel functionalities in React. It also updates the `@types/node` dependency from `20.10.3` to `20.10.4`. Additionally, there seems to be an indirect dependency update for `csstype` to `3.1.3` and `electron-to-chromium` to `1.4.608`.

The addition of `embla-carousel-react` includes corresponding updates to the lock file to ensure consistent installation of this package and its dependencies. The updates to `@types/node`, `csstype`, and `electron-to-chromium` potentially include minor bug fixes or enhancements that align with the node types definitions, CSS type definitions, and Electron-to-Chromium mappings respectively.

As these are non-breaking changes, they aim to enhance the functionality and developer experience without affecting existing codebases adversely.