style(css): standardize root font size and set base `body` font size
This commit introduces a standard font-size of 62.5% on the html element, which simplifies rem calculations to having 1rem equal 10px. Additionally, it establishes a base font-size of 1.6rem for the body element, effectively setting the default font-size to 16px when using the user-agent's default font settings. This can improve readability and makes setting font sizes more intuitive across the application.