feat(tag-component): create and export new Tag component
Created a new `Tag` component within the app, enabling the display of tag-like elements with custom styles and properties. The component is using the existing `Paragraph` component for text representation and utility function `cn` for conditional class name concatenation. The TagProps interface allows for passing a string `name` and additional classNames to the component.