feat(tag-mapper): add new TagMapper component
Add a new `TagMapper` component to map through an array of tags and render them as a list, utilizing generic typing for flexibility with different tag data structures. This includes the core component file as well as an index file to facilitate exports from the `tag-mapper` directory. The component accepts custom class names for styling individual tags and the wrapper element.