chore(dependencies): update ESLint and @types/react-dom packages
This commit updates the ESLint package from version 8.55.0 to 8.56.0 and the @types/react-dom package from version 18.2.17 to 18.2.18. The pnpm-lock.yaml file has been updated accordingly to reflect the new versions of these dependencies and their associated peer dependencies.

This update includes a minor bump for both packages, which are part of the development environment and should not affect production builds. It is important to ensure that the upgraded ESLint version does not introduce new linting issues by running the linter after this update.