chore(deps): bump @heroicons/react to 2.1.1 and tailwindcss to 3.3.7
This commit updates dependencies in the project, specifically increasing the version of `@heroicons/react` from 2.0.18 to 2.1.1 and upgrading `tailwindcss` from 3.3.6 to 3.3.7. Alongside these dependency version bumps, related integrity hashes have been updated in the `pnpm-lock.yaml` file. Additionally, a minor version update for `electron-to-chromium` which is a sub-dependency has also been included.