feat(dependencies): add @radix-ui/react-toggle and update tailwindcss to 3.4.0
This commit introduces a new dependency, @radix-ui/react-toggle at version 1.0.3, into the project's `package.json`. Additionally, it updates the version of `tailwindcss` from 3.3.7 to 3.4.0. The `pnpm-lock.yaml` file has also been updated to reflect these changes, ensuring that the package versions are locked and the integrity of dependencies is maintained.