chore(nvm): update Node version to use LTS
This commit updates the .nvmrc file to specify the use of the latest Long Term Support (LTS) version of Node.js instead of a hardcoded version number. This change ensures that the project will be using a version of Node.js that receives all critical bug fixes, security updates, and performance improvements while maintaining compatibility with the wider Node ecosystem. Using the LTS wildcard also reduces the maintenance overhead of updating the .nvmrc file when new Node.js versions are released.