refactor(ui): refactor WhiteButton to behave as a link
This commit refactors the `WhiteButton` component and introduces the `WhiteLink` component. The `WhiteLink` is styled similarly to the old button but is intended for external navigation, providing users with a more appropriate interaction when looking to book a call.

Additionally, this commit adds a link to the Calendly page for booking a call.